requirements.txt
cached-property==1.5.2
Cython==0.29.21
dataclasses==0.8
future==0.18.2
h5py==3.1.0
joblib==0.17.0
numpy==1.19.4
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3